gate/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.10.1
idna==2.8
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
gunicorn==19.9.0
gevent==1.4
django-cors-headers==3.1.0
